// extensions/minimum-subtotal/generated/input.ts
export type CartValidationsGenerateRunInput = { cart: { lines: Array<{ id: string }>, cost: { subtotalAmount: { amount: string, currencyCode: CurrencyCode } } }, validation: { config: { jsonValue: unknown } | null } };